elasticSearch 初步完成

+++ src/main/java/com/tj/demo/xunwu/repository/HouseRepository.java
@@ -17,7 +17,7 @@ public interface HouseRepository extends PagingAndSortingRepository<House,Intege
 
     @Modifying
     @Query("update House as house set house.status=:status where house.id=:id")
-    void updateStatus(@Param(value = "id") Long id,@Param(value="status") int status);
+    void updateStatus(@Param(value = "id") Integer id,@Param(value="status") int status);
 
     @Modifying
     @Query(" update House as house set house.watchTimes = house.watchTimes + 1 where house.id = :id ")
